# Conflicts:
#	LegacyCubiomes/utils/rng.hpp
#	programs/gen_chunk.cpp

+++ LegacyCubiomes/utils/rng.hpp
@@ -53,8 +53,10 @@ public:
         rng.setSeed(worldSeed);
         auto a = (i64) rng.nextLong();
         auto b = (i64) rng.nextLong();
-        a = (i64) (((a / 2) * 2) + 1);
-        b = (i64) (((b / 2) * 2) + 1);
+        // a = (i64) (((a / 2) * 2) + 1);
+        // b = (i64) (((b / 2) * 2) + 1);
+        a = ((a / 2) * 2) + 1;
+        b = ((b / 2) * 2) + 1;
         c_i64 decoratorSeed = (chunkX * a + chunkZ * b) ^ worldSeed;
         rng.setSeed(decoratorSeed);
         return rng;
